Event() - use unsigned short for keyCode constructor

+++ Rasterizer/Rasterizer/RasterizerState.hpp
@@ -16,7 +16,7 @@ struct RasterizerState {
         Event() {}
         Event(double time, Type type, float x, float y) : time(time), type(type), x(x), y(y), keyCode(0) {}
         Event(double time, Type type, Ra::Bounds b) : time(time), type(type), bounds(b) {}
-        Event(double time, Type type, int keyCode) : time(time), type(type), x(0.f), y(0.f), keyCode(keyCode) {}
+        Event(double time, Type type, unsigned short keyCode) : time(time), type(type), x(0.f), y(0.f), keyCode(keyCode) {}
         Event(double time, Type type, size_t flags) : time(time), type(type), x(0.f), y(0.f), keyCode(0), flags(flags) {}
         
         double time;
